Rename Params to Inbox. Add to env before calling handler.
